feat: check the directory, not just package.json
When checking for a local package, look at the whole directory including
the possibility of it being a monorepo. This is a prework for adding
license whitelisting.

Also add a missing `this.reset()` call when checking for a github PR.

And also fix a bug where the license was not checked for local
package.json files. Added a test case for that too.

Change-Id: Id655c7580d525b3db3e72259d069b99d17bfa8a4