feat: support local package whitelist
Packages listed in the whitelist are considered green. I was going to
support version spec as well, such as foo@^2.3.4, but that was trickier
to implement than I initially thought because comparing two version
ranges is tricky. So just whitelist using the package names.

This is the first step of related changes; to support package whitelist
for local repositories. Similar changes will be made for remote npm
packages and for github PRs.

For this, introduce a config file, js-green-licenses.json, in the same
directory as package.json.

Also rename the installed binary from `green-licenses` to `jsgl`.

Change-Id: Ie74d29154bb23d83fa381d18331b4552fe73db79